refactor: group statistics components
