+++ src/statistics/ItemReport.tsx
@@ -1,9 +1,9 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { invoke } from "@tauri-apps/api/core";
-import type { CatalogItem } from "./types/items";
-import type { SnapshotPoint, TrackedItem } from "./types/inventory";
-import { PREFERENCE_KEYS } from "./constants/preferences";
-import { TAURI_COMMANDS } from "./constants/tauri";
+import type { CatalogItem } from "../types/items";
+import type { SnapshotPoint, TrackedItem } from "../types/inventory";
+import { PREFERENCE_KEYS } from "../constants/preferences";
+import { TAURI_COMMANDS } from "../constants/tauri";
 import "./ItemReport.css";
 
 type Timeframe = "7" | "30" | "90" | "all";
